// src/types/TypeRutinasResponse.ts
// Generated by https://quicktype.io

export interface ResponseListaRutinas {
  ok: boolean;
  message: string;
  data: DatumRutinas[];
}

export interface DatumRutinas {
  id: string;
  nombre: string;
  createdAt: string;
  updatedAt: string;
  TBL_REL_CLIENTE_RUTINA: TblRelClienteRutina[];
  TBL_REL_RUTINA_EJERCICIO: TblRelRutinaEjercicio[];
}

export interface TblRelClienteRutina {
  id: string;
  id_cliente: string;
  id_rutina: string;
}

export interface TblRelRutinaEjercicio {
  id: string;
  id_rutina_entrenamiento: string;
  id_ejercicio_rutina: string;
  series: string;
  repeticiones: string;
  ejercicioRutina: EjercicioRutina;
  rutinaEntrenamiento: EjercicioRutina;
}

export interface EjercicioRutina {
  id: string;
  nombre: string;
  createdAt: string;
  updatedAt: string;
}